graphics.py is replaced by a module that also contains another module called modes; main.py now contains the User InstructionGroup and the MainWidget and imports all of the graphics from the graphics module
