+++ graphics/offbeat_spray.py
@@ -1,0 +1,64 @@
+import sys
+sys.path.append('..')
+from common.gfxutil import *
+
+from kivy.graphics.instructions import InstructionGroup
+from kivy.graphics import Color
+from kivy.graphics import PushMatrix, PopMatrix, Translate
+
+from random import random, choice
+
+import numpy as np
+
+
+class OffBeatSpray(InstructionGroup):
+    """Deprecated"""
+    def __init__(self, pos, colors, bound=50, dot_size=5, num_dots=10, animate=False):
+        """
+        If the player hits the spacebar off the beat, a spray of dots will be placed onscreen.
+        The colors map to the two most prominent pitch classes of that beat.
+
+        :param pos: (tuple) center position of spray in pixels, e.g. (1040, 980)
+        :param colors: (tuple) possible tuples for the color, e.g. ((255, 255, 255), (155, 155, 251))
+        :param bound: (int) radial bound on the spray
+        :param dot_size: (int) radius of the dots
+        :param num_dots: (int) number of dots in the spray
+        :param animate: (boolean) True if the dots will move around within the bound, False otherwise
+        """
+        super().__init__()
+        self.pos = np.array(pos)
+        self.colors = colors
+        self.bound = bound
+        self.rad = dot_size
+        self.num_dots = num_dots
+        self.animate = animate
+
+        # Time keeping for animation?
+        self.time = 0
+
+        # The entire spray should be able to move as one object
+        self.add(PushMatrix())
+        self.translate = Translate(*pos)
+        self.add(self.translate)
+
+        # For each dot in the spray, pick a random x, y coordinates using random radius, theta in ranges
+        # [0, self.bound], [0, 2pi] respectively
+        #       x = x_0 + r * cos(theta)
+        #       y = y_0 + r * sin(theta)
+        self.dots = [CEllipse(cpos=(self.pos[0] + self.bound * random() * np.cos(2 * np.pi * random()),
+                                    self.pos[1] + self.bound * random() * np.sin(2 * np.pi * random())),
+                              size=(self.rad, self.rad), segments=4) for i in range(self.num_dots)]
+
+        for dot in self.dots:
+            self.add(Color(*choice(self.colors)))
+            self.add(dot)
+
+        self.add(PopMatrix())
+
+    def on_update(self, time):
+        self.time = time
+        if self.animate:
+            for dot in self.dots:
+                dot.pos = (dot.pos[0] + choice((-1, 1)) * random() * 2,
+                           dot.pos[1] + choice((-1, 1)) * random() * 2)
+        self.translate.x = self.translate.x - 10
