+++ graphics/on_beat_bubble.py
@@ -1,0 +1,50 @@
+import sys
+sys.path.append('..')
+from common.gfxutil import *
+
+from kivy.graphics.instructions import InstructionGroup
+from kivy.graphics import Color
+
+
+class OnBeatBubble(InstructionGroup):
+    """Deprecated"""
+    def __init__(self, pos, radius, color, speed):
+        super().__init__()
+
+        self.original_pos = pos
+        self.radius = radius
+
+        # Color
+        self.color = Color(*color)
+        self.add(self.color)
+
+        # Shape
+        self.dot = CEllipse(cpos=pos, csize=(2 * radius, 2 * radius), sizesegments=20)
+        self.add(self.dot)
+
+        # Animation
+        self.beat_anim = None
+        self.pos_anim = KFAnim((0, pos[0]), (speed/pos[0], 0))
+        self.shrink = False
+
+        # Time keeping
+        self.time = 0
+        self.on_update(0)
+
+    def on_beat(self):
+        self.beat_anim = KFAnim((self.time, self.radius), (self.time+0.1, 2*self.radius))
+
+    def on_update(self, time):
+        self.dot.cpos = (self.dot.cpos[0] - 10, self.dot.cpos[1])
+
+        if self.beat_anim is not None:
+            radius = self.beat_anim.eval(self.time)
+            self.dot.csize = (2*radius, 2*radius)
+            if radius == 2*self.radius:
+                self.beat_anim = KFAnim((self.time, 2*self.radius), (self.time+0.1, self.radius))
+                self.shrink = True
+            if radius == self.radius and self.shrink:
+                self.beat_anim = None
+                self.shrink = False
+
+        self.time = time
