+++ main.py
@@ -1,12 +1,23 @@
 import sys
+import threading
 sys.path.append('..')
 
-from graphics import *
-from spotify import *
-import threading
-import time
+from common.core import *
+from common.gfxutil import *
 
-# https://developer.spotify.com/documentation/web-api/reference/tracks/get-audio-analysis/#timbre
+from kivy.graphics.instructions import InstructionGroup
+
+from random import random, choice, shuffle
+
+from graphics.background import AmbientBackgroundBlobs
+from graphics.modes.kaleidoscope import Kaleidoscope
+from graphics.modes.pulsing_bar import PulsingBar
+from graphics.modes.prism import Prism
+from graphics.modes.spectral_bars import SpectralBars
+from graphics.modes.tunnel import Tunnel
+from graphics.mode_transition import ModeTransition
+from graphics.progress_bar import ProgressBar
+from spotify import *
 
 
 global song_time
@@ -17,6 +28,163 @@ is_playing = False
 
 global song_name
 song_name = ""
+
+# https://developer.spotify.com/documentation/web-api/reference/tracks/get-audio-analysis/#timbre
+
+
+class User(InstructionGroup):
+    """ The User handles the audio analysis data and can call methods on the visualizer modes. """
+
+    def __init__(self, spotify_song, progress_bar):
+        super().__init__()
+
+        self.audio = spotify_song
+        self.progress_bar = progress_bar
+
+        num_sections = len(self.audio.get_current_track().get_sections())
+
+        # list of all background "shape" categories
+        self.backgrounds = ["a", "b", "c", "d"]
+
+        # the background for each section is random
+        self.section_backgrounds = [int(random() * len(self.backgrounds)) for _ in range(num_sections)]
+
+        # instance of current background
+        self.current_background = None
+
+        # list of all modes
+        self.modes = [PulsingBar, Tunnel, SpectralBars, Prism, Kaleidoscope]
+
+        # the mode for each section
+        self.section_modes = [int(random() * len(self.modes)) for i in
+                              range(len(self.audio.get_current_track().get_sections()))]
+        # self.section_modes = [4 for i in range(len(self.audio.get_current_track().get_sections()))]
+
+        # instance of current mode
+        self.current_mode = None
+
+        self.is_onbeat = False
+        self.last_beat = 0
+
+        self.in_transition = False
+        self.transition_time = 2000  # ms
+
+        # count the number of iterations that have been on beat
+        self.num_beats = 0
+
+        self.current_segment = 0
+        self.current_section = -1
+
+        # Time keeping
+        self.time = 0
+        self.on_update(0)
+
+        self.play = True
+
+    def play(self):
+        self.play = True
+
+    def pause(self):
+        self.play = False
+
+    def spacebar(self):
+        if self.play:
+            self.play = False
+        else:
+            self.play = True
+
+    def update_progress_bar(self, progress_bar):
+        self.progress_bar = progress_bar
+
+    def on_touch_move(self, touch):
+        if self.current_mode:
+            self.current_mode.on_touch_move(touch)
+
+    def on_update(self, time):
+        self.time = time
+
+        self.is_onbeat = self.audio.get_current_track().on_beat(self.time, 0.1)
+
+        section_index = self.audio.get_current_track().get_section_index(time)
+
+        if section_index == -1:
+            return
+
+        time_to_next = self.audio.get_current_track().get_time_to_next_section(time)
+
+        if time_to_next < self.transition_time / 1000 and not self.in_transition and time_to_next != -1:
+            self.in_transition = True
+
+            new_background = AmbientBackgroundBlobs(shape=choice(self.backgrounds))
+            if self.current_background:
+                self.remove(self.current_background)
+            self.add(new_background)
+
+            self.current_background = new_background
+
+            color = self.progress_bar.get_section_color(section_index + 1)
+            shuffle(self.modes)
+            new_mode = ModeTransition(color, self.modes[0], self.modes[1], self.time, self.transition_time)
+
+            if self.current_mode:
+                self.remove(self.current_mode)
+            self.add(new_mode)
+
+            self.current_mode = new_mode
+            self.current_section = section_index + 1
+
+        elif section_index != self.current_section and not self.in_transition and section_index != -1:
+            self.in_transition = True
+
+            new_background = AmbientBackgroundBlobs(shape=choice(self.backgrounds))
+            if self.current_background:
+                self.remove(self.current_background)
+            self.add(new_background)
+
+            self.current_background = new_background
+
+            color = self.progress_bar.get_section_color(section_index)
+            shuffle(self.modes)
+            new_mode = ModeTransition(color, self.modes[0], self.modes[1], self.time, self.transition_time)
+
+            if self.current_mode:
+                self.remove(self.current_mode)
+            self.add(new_mode)
+
+            self.current_mode = new_mode
+            self.current_section = section_index
+
+        self.in_transition = self.current_mode.in_transition(self.time, self.transition_time)
+
+        # if section_index != self.current_section:
+        #     # new section
+        #     # print(section_index)
+        #     new_mode = self.modes[self.section_modes[section_index]](self.progress_bar.get_section_color(section_index))
+        #     if self.current_mode:
+        #         self.remove(self.current_mode)
+        #     self.add(new_mode)
+
+        #     self.current_mode = new_mode
+        #     self.current_section = section_index
+
+        if self.is_onbeat:
+            self.num_beats += 1
+        else:
+            self.num_beats = 0
+
+        if self.is_onbeat:
+            self.current_mode.on_beat()
+
+        segment_index = self.audio.get_current_track().get_segment_index(time)
+        if segment_index != self.current_segment:
+            self.current_segment = segment_index
+            data = self.audio.get_current_track().get_segments_data()[self.current_segment]
+            self.current_mode.on_segment(data)
+            self.progress_bar.on_segment(data)
+
+        self.current_mode.on_update(self.time)
+        self.current_background.on_update(self.time)
+
 
 class MainWidget(BaseWidget):
     def __init__(self):
@@ -130,6 +298,7 @@ class MainWidget(BaseWidget):
         self.progress_bar.on_update(self.progress)
 
         # print(self.audio.get_current_track().get_section_index(self.time))
+
 
 def song_label(font_filepath='res/font/CabinCondensed-Regular.ttf', font_size='20sp'):
     l = Label(text = "text", valign='top', halign='right', font_size=font_size,
